# src/routes/categoriesRoutes.js
/**
 * @swagger
 * tags:
 *   name: Categories
 *   description: Rotas para gerenciar categorias
 *
 * /categories:
 *   get:
 *     summary: Lista todas as categorias
 *     tags: [Categories]
 *     responses:
 *       200:
 *         description: Lista de categorias retornada com sucesso
 *   post:
 *     summary: Cria uma nova categoria
 *     tags: [Categories]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               name:
 *                 type: string
 *     responses:
 *       201:
 *         description: Categoria criada com sucesso
 *
 * /categories/{id}:
 *   get:
 *     summary: Busca uma categoria pelo ID
 *     tags: [Categories]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *     responses:
 *       200:
 *         description: Categoria encontrada
 *       404:
 *         description: Categoria não encontrada
 *   put:
 *     summary: Atualiza uma categoria pelo ID
 *     tags: [Categories]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               name:
 *                 type: string
 *     responses:
 *       200:
 *         description: Categoria atualizada com sucesso
 *   delete:
 *     summary: Remove uma categoria pelo ID
 *     tags: [Categories]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *     responses:
 *       204:
 *         description: Categoria removida com sucesso
 */


const express = require('express');
const router = express.Router();
const categoriesControllers = require('../controllers/categoriesControllers.js');
const Authenticated = require('../middlewares/authenticateToken');

// Middleware to authenticate routes
router.use(Authenticated);

// Route to get all categories

router.get('/', categoriesControllers.findAllCategories);
router.get('/:id', categoriesControllers.findCategoryById);
router.post('/', categoriesControllers.createCategory);
router.put('/:id', categoriesControllers.updateCategory);
router.delete('/:id', categoriesControllers.deleteCategory);
router.get('/products/:id', categoriesControllers.findProductsByCategoryId);

module.exports = router;