# src/routes/usersRoutes.js
/**
 * @swagger
 * tags:
 *   name: Users
 *   description: Rotas para gerenciar usuários
 *
 * /users:
 *   get:
 *     summary: Lista todos os usuários
 *     tags: [Users]
 *     responses:
 *       200:
 *         description: Lista de usuários retornada com sucesso
 *   post:
 *     summary: Cria um novo usuário
 *     tags: [Users]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               username:
 *                 type: string
 *               email:
 *                 type: string
 *               password:
 *                 type: string
 *     responses:
 *       201:
 *         description: Usuário criado com sucesso
 *
 * /users/{id}:
 *   get:
 *     summary: Busca um usuário pelo ID
 *     tags: [Users]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *     responses:
 *       200:
 *         description: Usuário encontrado
 *       404:
 *         description: Usuário não encontrado
 *   put:
 *     summary: Atualiza um usuário pelo ID
 *     tags: [Users]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               username:
 *                 type: string
 *               email:
 *                 type: string
 *     responses:
 *       200:
 *         description: Usuário atualizado com sucesso
 *   delete:
 *     summary: Remove um usuário pelo ID
 *     tags: [Users]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *     responses:
 *       204:
 *         description: Usuário removido com sucesso
 */

const express = require('express');
const router = express.Router();

const UserController = require('../controllers/usersControllers');  
const validateLogin = require('../middlewares/validateLogin');
const authenticateToken = require('../middlewares/authenticateToken');

router.post('/login', validateLogin, UserController.login);
router.put('/:id', authenticateToken, UserController.updateUser);
router.post('/', UserController.createUser);
router.get('/', UserController.findAllUsers);
router.get('/:id', UserController.findUserById);
router.delete('/:id', authenticateToken, UserController.deleteUser);
router.get('/:id/orders', authenticateToken, UserController.findOrdersByUserId);

module.exports = router;